requirements.txt
click==8.1.8
jinja2==3.1.6
lxml==5.3.0
lxml-stubs==0.5.1
python-dotenv==1.0.1
PyYAML==6.0.2
types-PyYAML==6.0.12.20241230
docstring-parser==0.16
litellm==1.81.8
geopandas==1.1.1
# sandbox
ipykernel==6.29.5
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.22
faster-whisper==1.0.3
# for testing
pytest==9.0.2
